+++ resources/js/g/admissao/import/app.js
@@ -1,0 +1,12 @@
+import { createApp } from 'vue'
+import { registerGlobals } from '../../../registerGlobals'
+import ImportacaoAdmissoes from '../../../components/admissao/import/ImportacaoAdmissoes.vue'
+
+const app = createApp({
+    components: {
+        ImportacaoAdmissoes
+    }
+})
+
+registerGlobals(app)
+app.mount('#app')